brand-image-agent/pyproject.toml
[project]
name = "brand-image-agent"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "google-adk>=1.27.2",
    "google-genai>=1.68.0",
    "onnxruntime>=1.24.4",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
    "rembg>=2.0.73",
    "sqlalchemy>=2.0.48",
]
